// resources/js/api/generated/models/operationLogDetail.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { OperationLogMessage } from './operationLogMessage';
import type { OperationLogResponse } from './operationLogResponse';

export interface OperationLogDetail {
  log: OperationLogResponse;
  messages: OperationLogMessage[];
}